sibu 1.0.7 added form input helper

+++ app/models/sibu/dynamic_style.rb
@@ -1,31 +1,36 @@
 module Sibu
   class DynamicStyle
 
-    attr_reader :site, :body, :env, :filename, :scss_file
+    attr_reader :site, :body, :env, :filename, :scss_file, :styles_changed
 
     def initialize(site_id)
       @site = Sibu::Site.find(site_id)
-      @filename = "#{site_id}_#{Time.current.to_i}"
-      @scss_file = File.new(scss_file_path, 'w')
-      @body = ERB.new(File.read(template_file_path)).result(binding)
-      @env = Rails.application.assets
+      @styles_changed = styles_changed?(@site.style_url)
+      if @styles_changed
+        @filename = "#{site_id}_#{Time.current.to_i}"
+        @scss_file = File.new(scss_file_path, 'w')
+        @body = ERB.new(File.read(template_file_path)).result(binding)
+        @env = Rails.application.assets
+      end
     end
 
     def compile
-      find_or_create_scss
+      if @styles_changed
+        find_or_create_scss
 
-      begin
-        scss_file.write generate_css
-        scss_file.flush
-        scss_file.close
-        css_file_path = scss_file_path.gsub('scss', 'css')
-        File.rename(scss_file_path, css_file_path)
-        site.update(style: File.new(css_file_path))
-      rescue Exception => ex
-          Rails.logger.error(ex)
-      ensure
-        File.delete(scss_file_path) if scss_file_path && File.exist?(scss_file_path)
-        File.delete(css_file_path) if css_file_path && File.exist?(css_file_path)
+        begin
+          scss_file.write generate_css
+          scss_file.flush
+          scss_file.close
+          css_file_path = scss_file_path.gsub('scss', 'css')
+          File.rename(scss_file_path, css_file_path)
+          site.update(style: File.new(css_file_path))
+        rescue Exception => ex
+            Rails.logger.error(ex)
+        ensure
+          File.delete(scss_file_path) if scss_file_path && File.exist?(scss_file_path)
+          File.delete(css_file_path) if css_file_path && File.exist?(css_file_path)
+        end
       end
     end
 
@@ -72,6 +77,17 @@ module Sibu
         asset = Sprockets::UnloadedAsset.new(uri, env)
         env.load(asset.uri).source
       end
+    end
+
+    def styles_changed?(site_style_url)
+      templates_styles_path = File.join(Rails.root, 'app', 'assets', 'stylesheets', 'templates', '**/*')
+      style_updated_at = File.new(File.join(Rails.root, 'public', site_style_url)).mtime
+      Dir.glob(templates_styles_path) do |style_file|
+        if File.new(style_file).mtime > style_updated_at
+          return true
+        end
+      end
+      false
     end
   end
 end
